# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: TAIFA Platform API
  version: 1.0.0
  description: Authoritative TAIFA Platform APIs for identity-bound access, payments
    over the double-entry ledger, enterprise financial operations, commerce and national
    mobility.
paths:
...
  /api/v1/trips/emergency/dispatch:
    post:
      operationId: api_v1_trips_emergency_dispatch_create
      tags:
      - mobility-emergency
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
  /api/v1/trips/intercity/corridors:
    get:
      operationId: api_v1_trips_intercity_corridors_retrieve
      tags:
      - mobility-intercity
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
    post:
      operationId: api_v1_trips_intercity_corridors_create
      tags:
      - mobility-intercity
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
  /api/v1/trips/logistics/shipments:
    post:
      operationId: api_v1_trips_logistics_shipments_create
      tags:
      - mobility-logistics
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
  /api/v1/trips/national/analytics:
    get:
      operationId: api_v1_trips_national_analytics_retrieve
      tags:
      - mobility-national
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
  /api/v1/trips/national/command-center:
    get:
      operationId: api_v1_trips_national_command_center_retrieve
      tags:
      - mobility-national
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
  /api/v1/trips/national/map:
    get:
      operationId: api_v1_trips_national_map_retrieve
      parameters:
      - in: query
        name: region
        schema:
          type: string
      tags:
      - mobility-national
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
  /api/v1/trips/public-transit/routes:
    get:
      operationId: api_v1_trips_public_transit_routes_retrieve
      tags:
      - mobility-public-transit
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
    post:
      operationId: api_v1_trips_public_transit_routes_create
      tags:
      - mobility-public-transit
      security:
      - DeviceToken: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    DeviceToken:
      type: http
      scheme: bearer
      description: 'Device-bound token from `POST /auth/device/register`. Send it
        as `Authorization: Bearer <token>` together with the `X-Device-Id` header.'
servers:
- url: /
  description: current host
...